requirements.txt
python-openhab-rest-client